[Feat] #44 - SnowmanType, SnowmanItemType Shared 모듈로 이동
