+++ Projects/DesignSystem/Sources/Snowman/Item/SnowmanItemImageType.swift
@@ -8,40 +8,6 @@
 
 import SwiftUI
 
-public enum SnowmanItemType: Int {
-    // Priority 1 : Eye, Mouth
-    case button
-    case stones
-    case saw
-    
-    // Priority 10 : Nose
-    case carrot
-    case radish
-    
-    // Priority 50
-    case gloves
-    case cardigan
-    case branch
-    
-    case earMuff
-    case airPodMax1
-    case airPodMax2
-    case airPodMax3
-    
-    // Priority 70
-    case muffler1
-    case muffler2
-    
-    // Priority 200
-    case boots
-    
-    // Priority None
-    case bikini
-    case can
-    case santaHat
-    case sunglass
-}
-
 public enum SnowmanItemImageType {
     // Priority 1 : Eye, Mouth
     case button
