--- Projects/Domain/Sources/Domain.swift
@@ -1,1 +1,0 @@
-import SwiftUI
