+++ Projects/Shared/Sources/Snowman/SnowmanType.swift
@@ -1,0 +1,14 @@
+//
+//  SnowmanType.swift
+//  Shared_ios
+//
+//  Created by 김윤서 on 2023/07/31.
+//  Copyright © 2023 EarthIsRound. All rights reserved.
+//
+
+public enum SnowmanType {
+    case largeHeadLargeBody
+    case largeHeadSmallBody
+    case smallHeadLargeBody
+    case smallHeadSmallBody
+}
